+++ actionpack/test/controller/show_exceptions_test.rb
@@ -3,6 +3,7 @@ require 'abstract_unit'
 module ShowExceptions
   class ShowExceptionsController < ActionController::Base
     use ActionDispatch::ShowExceptions
+    use ActionDispatch::DebugExceptions
 
     def boom
       raise 'boom!'